--- a/src/db/Table.php
+++ b/src/db/Table.php
@@ -60,6 +60,8 @@ abstract class Table
     public const GQLTOKENS = '{{%gqltokens}}';
     public const INFO = '{{%info}}';
     public const MATRIXBLOCKS = '{{%matrixblocks}}';
+    /** @since 4.0.0 */
+    public const MATRIXBLOCKS_OWNERS = '{{%matrixblocks_owners}}';
     public const MATRIXBLOCKTYPES = '{{%matrixblocktypes}}';
     public const MIGRATIONS = '{{%migrations}}';
     /** @since 3.4.0 */