--- a/src/fields/Matrix.php
+++ b/src/fields/Matrix.php
@@ -578,7 +578,11 @@ class Matrix extends Field implements EagerLoadingFieldInterface, GqlInlineFragm
         $existsQuery = (new Query())
             ->from(["matrixblocks_$ns" => DbTable::MATRIXBLOCKS])
             ->innerJoin(["elements_$ns" => DbTable::ELEMENTS], "[[elements_$ns.id]] = [[matrixblocks_$ns.id]]")
-            ->where("[[matrixblocks_$ns.ownerId]] = [[elements.id]]")
+            ->innerJoin(["matrixblocks_owners_$ns" => DbTable::MATRIXBLOCKS_OWNERS], [
+                'and',
+                "[[matrixblocks_owners_$ns.blockId]] = [[elements_$ns.id]]",
+                "[[matrixblocks_owners_$ns.ownerId]] = [[elements.id]]"
+            ])
             ->andWhere([
                 "matrixblocks_$ns.fieldId" => $this->id,
                 "elements_$ns.enabled" => true,
@@ -852,13 +856,18 @@ class Matrix extends Field implements EagerLoadingFieldInterface, GqlInlineFragm
 
         // Return any relation data on these elements, defined with this field
         $map = (new Query())
-            ->select(['ownerId as source', 'id as target'])
-            ->from([DbTable::MATRIXBLOCKS])
-            ->where([
-                'fieldId' => $this->id,
-                'ownerId' => $sourceElementIds,
+            ->select([
+                'source' => 'matrixblocks_owners.ownerId',
+                'target' => 'matrixblocks.id',
             ])
-            ->orderBy(['sortOrder' => SORT_ASC])
+            ->from(['matrixblocks' => DbTable::MATRIXBLOCKS])
+            ->innerJoin(['matrixblocks_owners' => DbTable::MATRIXBLOCKS_OWNERS], [
+                'and',
+                '[[matrixblocks_owners.blockId]] = [[matrixblocks.id]]',
+                ['matrixblocks_owners.ownerId' => $sourceElementIds],
+            ])
+            ->where(['matrixblocks.fieldId' => $this->id])
+            ->orderBy(['matrixblocks_owners.sortOrder' => SORT_ASC])
             ->all();
 
         return [
@@ -1011,7 +1020,12 @@ class Matrix extends Field implements EagerLoadingFieldInterface, GqlInlineFragm
         $resetValue = false;
 
         if ($element->duplicateOf !== null) {
-            $matrixService->duplicateBlocks($this, $element->duplicateOf, $element, true);
+            // If this is a draft, just duplicate the relations
+            if ($element->getIsDraft()) {
+                $matrixService->duplicateOwnership($this, $element->duplicateOf, $element);
+            } else {
+                $matrixService->duplicateBlocks($this, $element->duplicateOf, $element, true);
+            }
             $resetValue = true;
         } else if ($element->isFieldDirty($this->handle) || !empty($element->newSiteIds)) {
             $matrixService->saveField($this, $element);
@@ -1040,13 +1054,13 @@ class Matrix extends Field implements EagerLoadingFieldInterface, GqlInlineFragm
             return false;
         }
 
-        // Delete any Matrix blocks that belong to this element(s)
+        // Delete any Matrix blocks that primarily belong to this element
         foreach (Craft::$app->getSites()->getAllSiteIds() as $siteId) {
             $elementsService = Craft::$app->getElements();
             $matrixBlocks = MatrixBlock::find()
                 ->status(null)
                 ->siteId($siteId)
-                ->ownerId($element->id)
+                ->primaryOwnerId($element->id)
                 ->all();
 
             foreach ($matrixBlocks as $matrixBlock) {
@@ -1069,7 +1083,7 @@ class Matrix extends Field implements EagerLoadingFieldInterface, GqlInlineFragm
             $blocks = MatrixBlock::find()
                 ->status(null)
                 ->siteId($siteInfo['siteId'])
-                ->ownerId($element->id)
+                ->primaryOwnerId($element->id)
                 ->trashed()
                 ->andWhere(['matrixblocks.deletedWithOwner' => true])
                 ->all();
@@ -1250,7 +1264,22 @@ class Matrix extends Field implements EagerLoadingFieldInterface, GqlInlineFragm
             // Existing block?
             if (isset($oldBlocksById[$blockId])) {
                 $block = $oldBlocksById[$blockId];
-                $block->dirty = !empty($blockData);
+                $dirty = !empty($blockData);
+
+                // Is this a derivative element, and does the block primarily belong to the canonical?
+                if ($dirty && $element->getIsDerivative() && $block->primaryOwnerId === $element->getCanonicalId()) {
+                    // Duplicate it as a draft. (We'll drop its draft status from `Matrix::saveField()`.)
+                    $block = Craft::$app->getDrafts()->createDraft($block, Craft::$app->getUser()->getId(), null, null, [
+                        'canonicalId' => $block->id,
+                        'primaryOwnerId' => $element->id,
+                        'owner' => $element,
+                        'siteId' => $element->siteId,
+                        'propagating' => false,
+                        'markAsSaved' => false,
+                    ]);
+                }
+
+                $block->dirty = $dirty;
             } else {
                 // Make sure it's a valid block type
                 if (!isset($blockData['type']) || !isset($blockTypes[$blockData['type']])) {
@@ -1259,7 +1288,7 @@ class Matrix extends Field implements EagerLoadingFieldInterface, GqlInlineFragm
                 $block = new MatrixBlock();
                 $block->fieldId = $this->id;
                 $block->typeId = $blockTypes[$blockData['type']]->id;
-                $block->ownerId = $element->id;
+                $block->primaryOwnerId = $element->id;
                 $block->siteId = $element->siteId;
 
                 // Preserve the collapsed state, which the browser can't remember on its own for new blocks